requirements.txt
django==2.1
django-debug-toolbar
Pillow
